refactor(metrics): re-home metrics onto aragora.evaluation with shims (VAL-P4A-012)
Move the VIAH, Manifold Brier, and capability-checkpoint originals from
aragora/metrics/ to aragora/evaluation/ and leave aragora/metrics/*.py as
DeprecationWarning re-export shims so the legacy import paths keep working.
